Move OpenGL objects to View
