+++ src/Model/GameObject/Item.hpp
@@ -3,9 +3,9 @@
 #include <string>
 
 #include "GameObject.hpp"
-#include "Model.hpp"
-#include "Texture.hpp"
-#include "Material.hpp"
+#include "View/Renderer/OpenGL/Objects/Model.hpp"
+#include "View/Renderer/OpenGL/Objects/Texture.hpp"
+#include "View/Renderer/OpenGL/Objects/Material.hpp"
 
 /**
  * @class Item
