+++ src/Model/GameObject/PhysicsObject.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include "BodyRigid.hpp"
-#include "Model.hpp"
-#include "Texture.hpp"
-#include "Material.hpp"
+#include "View/Renderer/OpenGL/Objects/Model.hpp"
+#include "View/Renderer/OpenGL/Objects/Texture.hpp"
+#include "View/Renderer/OpenGL/Objects/Material.hpp"
 
 class PhysicsObject : public BodyRigid {
 public:
