+++ src/View/Renderer/OpenGL/Objects/OmniShadowMap.hpp
