+++ src/View/Renderer/OpenGL/Objects/PointLight.hpp
