+++ src/View/Renderer/OpenGL/Objects/Skybox.hpp
