+++ src/View/Renderer/OpenGL/Objects/SpotLight.cpp
