+++ src/View/Renderer/OpenGL/Objects/SpotLight.hpp
